requirements.txt
deltas
mediawiki-utilities
pymongo
pyyaml
